refactor: apiClientのディレクトリを修正
