+++ hooks/useDeletePost.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query'
-import { deletePost } from '../app/client/deletePost'
+import { deletePost } from '../app/client/microcms/post/deletePost'
 
 export const useDeletePost = () =>
   useMutation(async (contentId: string) => await deletePost(contentId))
